tanstack query migration

+++ api/src/team/team.service.ts
@@ -1,18 +1,20 @@
-import {
-  BadRequestException,
-  forwardRef,
-  Inject,
-  Injectable,
-  Logger,
-} from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
+import { forwardRef, Inject, Injectable, Logger } from "@nestjs/common";
+import { InjectDataSource, InjectRepository } from "@nestjs/typeorm";
 import { TeamEntity } from "./entities/team.entity";
-import { Repository } from "typeorm";
+import {
+  DataSource,
+  EntityManager,
+  IsNull,
+  LessThanOrEqual,
+  Repository,
+} from "typeorm";
 import { GithubApiService } from "../github-api/github-api.service";
 import { EventService } from "../event/event.service";
 import { UserService } from "../user/user.service";
 import { FindOptionsRelations } from "typeorm/find-options/FindOptionsRelations";
 import { MatchService } from "../match/match.service";
+import { Cron, CronExpression } from "@nestjs/schedule";
+import { LockKeys } from "../constants";
 
 @Injectable()
 export class TeamService {
@@ -25,9 +27,49 @@ export class TeamService {
     private readonly userService: UserService,
     @Inject(forwardRef(() => MatchService))
     private readonly matchService: MatchService,
+    @InjectDataSource()
+    private readonly dataSource: DataSource,
   ) {}
 
   logger = new Logger("TeamService");
+
+  @Cron(CronExpression.EVERY_MINUTE)
+  async autoCreateRepos() {
+    const lockKey = LockKeys.CREATE_TEAM_REPOS;
+    const queryRunner = this.dataSource.createQueryRunner();
+    await queryRunner.connect();
+
+    try {
+      const gotLock = await queryRunner.query(
+        "SELECT pg_try_advisory_lock($1)",
+        [lockKey],
+      );
+
+      if (gotLock[0].pg_try_advisory_lock) {
+        try {
+          const teams = await this.teamRepository.findBy({
+            startedRepoCreationAt: IsNull(),
+            event: {
+              startDate: LessThanOrEqual(new Date()),
+            },
+          });
+          for (const team of teams) {
+            this.logger.log(
+              `Starting repo creation for team ${team.name} as its event has started.`,
+            );
+
+            await this.dataSource.transaction(async (entityManager) => {
+              await this.createTeamRepository(team.id, entityManager);
+            });
+          }
+        } finally {
+          await queryRunner.query("SELECT pg_advisory_unlock($1)", [lockKey]);
+        }
+      }
+    } finally {
+      await queryRunner.release();
+    }
+  }
 
   getTeamById(
     id: string,
@@ -91,42 +133,66 @@ export class TeamService {
     });
   }
 
-  async createTeam(name: string, userId: string, eventId: string) {
-    const event = await this.eventService.getEventById(eventId);
-    const user = await this.userService.getUserById(userId);
-    const team = await this.teamRepository.save({
-      name,
-      event: { id: eventId },
-      users: [{ id: userId }],
+  async createTeamRepository(teamId: string, entityManager: EntityManager) {
+    const teamRepository = entityManager.getRepository(TeamEntity);
+    const team = await teamRepository.findOneOrFail({
+      where: {
+        startedRepoCreationAt: IsNull(),
+        id: teamId,
+      },
+      relations: {
+        users: true,
+        event: true,
+      },
     });
 
-    const repoName = event.name + "-" + name + "-" + team.id;
-    try {
-      await this.githubApiService.createTeamRepository(
-        repoName,
-        team.name,
-        user.username,
-        user.githubAccessToken,
-        event.githubOrg,
-        event.githubOrgSecret,
-        team.id,
-        event.monorepoUrl,
-        event.monorepoVersion,
-        event.myCoreBotDockerImage,
-        event.visualizerDockerImage,
-        event.id,
+    if (!team.event) {
+      this.logger.error(
+        `While creating repo for team ${teamId}, event was not found.`,
       );
-
-      await this.teamRepository.save(team);
-    } catch (e) {
-      this.logger.error(`Failed to create repository for team ${team.id}`, e);
-      await this.teamRepository.delete(team.id);
-      throw new BadRequestException(
-        "Failed to create repository for team. Please try again later.",
-      );
+      return;
     }
 
-    return team;
+    const repoName = team.event.name + "-" + team.name + "-" + team.id;
+
+    await this.githubApiService.createTeamRepository(
+      repoName,
+      team.name,
+      team.users.map((user) => ({
+        username: user.username,
+        githubAccessToken: user.githubAccessToken,
+      })),
+      team.event.githubOrg,
+      team.event.githubOrgSecret,
+      team.id,
+      team.event.monorepoUrl,
+      team.event.monorepoVersion,
+      team.event.myCoreBotDockerImage,
+      team.event.visualizerDockerImage,
+      team.event.id,
+    );
+
+    await teamRepository.update(teamId, {
+      repo: repoName,
+      startedRepoCreationAt: new Date(),
+    });
+  }
+
+  async createTeam(name: string, userId: string, eventId: string) {
+    return await this.dataSource.transaction(async (entityManager) => {
+      const teamRepository = entityManager.getRepository(TeamEntity);
+
+      const newTeam = await teamRepository.save({
+        name,
+        event: { id: eventId },
+        users: [{ id: userId }],
+      });
+
+      if (await this.eventService.hasEventStarted(eventId))
+        await this.createTeamRepository(newTeam.id, entityManager);
+
+      return newTeam;
+    });
   }
 
   async deleteTeam(teamId: string) {
@@ -151,7 +217,7 @@ export class TeamService {
     });
     const user = await this.userService.getUserById(userId);
 
-    if (team.users.length > 1) {
+    if (team.users.length > 1 && team.repo) {
       await this.githubApiService.removeUserFromRepository(
         team.repo,
         user.username,
@@ -222,13 +288,14 @@ export class TeamService {
     });
     const user = await this.userService.getUserById(userId);
 
-    await this.githubApiService.addUserToRepository(
-      team.repo,
-      user.username,
-      team.event.githubOrg,
-      team.event.githubOrgSecret,
-      user.githubAccessToken,
-    );
+    if (team.repo)
+      await this.githubApiService.addUserToRepository(
+        team.repo,
+        user.username,
+        team.event.githubOrg,
+        team.event.githubOrgSecret,
+        user.githubAccessToken,
+      );
 
     await this.teamRepository
       .createQueryBuilder()
